feat: Upgrade `useCachedNotification` handling
